+++ src/pages/PokedexPage/PokedexPage.jsx
@@ -1,9 +1,12 @@
 import { ConteinerPokedex, H1 } from "./styled";
 import { PokemonCard } from "../../components/pokemonCard/PokemonCard";
+import { useContext } from "react";
+import GlobalContext from "../../context/GlobalContext";
 
-export const PokedexPage = ({ pokedex, deletPokemon,setDetail }) => {
+export const PokedexPage = () => {
   
-
+  const context = useContext(GlobalContext)
+  const { pokedex } = context;
   return (
     <ConteinerPokedex>
       <H1>Meus pokémons</H1>
@@ -12,8 +15,6 @@ export const PokedexPage = ({ pokedex, deletPokemon,setDetail }) => {
           <PokemonCard
             key={index}
             pokemon={pokemon}
-            deletPokemon={deletPokemon}
-            setDetail={setDetail}
             btn = {'delet'}
           />
         );
